requirements.txt
python-telegram-bot==22.2
firebase-admin
